requirements.txt
Django
pycountry
django-datatables-view
djangorestframework
drf-form-bootstrap-4
django-crispy-forms
